+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.2
  * Make macro-generated types (*RPCClient, *RPCServer) public
  * Split Transport into ClientTransport and ServerTransport
+ * Pass state to rx_response